Added istanbul as a codecov tool.

+++ src/index.js
@@ -7,8 +7,9 @@ const Hoek = require('hoek');
 const fs = require('fs');
 const pool = require('src/middleware/db').pool;
 const routes = require('src/middleware/route_crawler')('./src/routes');
+const process = require('process');
 
-const server = new Hapi.Server();
+const server = module.exports = new Hapi.Server();
 
 server.connection({
   port: 3000,
@@ -17,34 +18,37 @@ server.connection({
 
 server.route(routes);
 
-server.register([require('vision'), require('inert'), { register: require('lout') }, { register: require('tv'), options: { endpoint: '/tv'} }], function (err) {
-  Hoek.assert(!err, err);
-
-  server.start(function (err) {
+/* istanbul ignore if  */
+if (process.env['NODE_ENV'] !== 'test') {
+  server.register([require('vision'), require('inert'), { register: require('lout') }, { register: require('tv'), options: { endpoint: '/tv'} }], function (err) {
     Hoek.assert(!err, err);
 
-    console.log('info', 'Server is starting');
+    server.start(function (err) {
+      Hoek.assert(!err, err);
 
-    pool.connect(function (err, client, done) {
-      client.query('DROP SCHEMA public CASCADE', (err, res) => {
-        console.log('info', 'Dropped schema public.');
+      console.log('info', 'Server is starting');
 
-        client.query('CREATE SCHEMA public', (err, res) => {
-          Hoek.assert(!err, err);
+      pool.connect(function (err, client, done) {
+        client.query('DROP SCHEMA public CASCADE', (err, res) => {
+          console.log('info', 'Dropped schema public.');
 
-          console.log('info', 'Created schema public.');
-
-          const query = fs.readFileSync('src/sql/init.sql').toString();
-
-          client.query(query, (err, res) => {
+          client.query('CREATE SCHEMA public', (err, res) => {
             Hoek.assert(!err, err);
 
-            console.log('info', 'Server running at: ' + server.info.uri);
+            console.log('info', 'Created schema public.');
 
-            done();
+            const query = fs.readFileSync('src/sql/init.sql').toString();
+
+            client.query(query, (err, res) => {
+              Hoek.assert(!err, err);
+
+              console.log('info', 'Server running at: ' + server.info.uri);
+
+              done();
+            });
           });
         });
       });
     });
   });
-});
+}
